@@ -50,8 +50,11 @@ type AccountParams struct {
 // TransactionParams allows the client to filter transactions
 type TransactionParams struct {
 	PaginationParams
-	Height uint64 `json:"height,omitempty"`
-	Type   string `json:"type,omitempty"`
+	Hash    string `json:"hash,omitempty"`
+	Height  uint64 `json:"height,omitempty"`
+	Type    string `json:"type,omitempty"`
+	Subtype string `json:"subtype,omitempty"`
+	Signer  string `json:"signer,omitempty"`
 }
 
 // BlockParams allows the client to filter blocks
@@ -275,8 +278,8 @@ type TransactionReference struct {
 
 // TransactionsList is used to return a paginated list to the client
 type TransactionsList struct {
-	Transactions []*indexertypes.Transaction `json:"transactions"`
-	Pagination   *Pagination                 `json:"pagination"`
+	Transactions []*indexertypes.TransactionMetadata `json:"transactions"`
+	Pagination   *Pagination                         `json:"pagination"`
 }
 
 // FeesList is used to return a paginated list to the client
@@ -292,9 +295,9 @@ type TransfersList struct {
 }
 
 type GenericTransactionWithInfo struct {
-	TxContent json.RawMessage          `json:"tx"`
-	TxInfo    indexertypes.Transaction `json:"txInfo"`
-	Signature types.HexBytes           `json:"signature"`
+	TxContent json.RawMessage           `json:"tx"`
+	TxInfo    *indexertypes.Transaction `json:"txInfo"`
+	Signature types.HexBytes            `json:"signature"`
 }
 
 type ChainInfo struct {
@@ -444,8 +447,9 @@ func CensusTypeToOrigin(ctype CensusTypeDescription) (models.CensusOrigin, []byt
 }
 
 type Block struct {
-	comettypes.Block `json:",inline"`
-	Hash             types.HexBytes `json:"hash" `
+	comettypes.Header `json:"header"`
+	Hash              types.HexBytes `json:"hash" `
+	TxCount           int64          `json:"txCount"`
 }
 
 // BlockList is used to return a paginated list to the client
