@@ -2,6 +2,7 @@ package indexer
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 	"math/big"
 	"sync"
@@ -14,6 +15,7 @@ import (
 	"go.vocdoni.io/dvote/test/testcommon/testutil"
 	"go.vocdoni.io/dvote/util"
 	"go.vocdoni.io/dvote/vochain"
+	indexerdb "go.vocdoni.io/dvote/vochain/indexer/db"
 	"go.vocdoni.io/dvote/vochain/state"
 	"go.vocdoni.io/dvote/vochain/transaction/vochaintx"
 	"go.vocdoni.io/proto/build/go/models"
@@ -85,6 +87,7 @@ func BenchmarkIndexer(b *testing.B) {
 				tx := &vochaintx.Tx{
 					TxID:        rnd.Random32(),
 					TxModelType: "vote",
+					Tx:          &models.Tx{Payload: &models.Tx_Vote{}},
 				}
 				idx.OnNewTx(tx, height, txBlockIndex)
 				curTxs = append(curTxs, tx)
@@ -112,7 +115,7 @@ func BenchmarkIndexer(b *testing.B) {
 						qt.Check(b, bytes.Equal(voteRef.Meta.TxHash, tx.TxID[:]), qt.IsTrue)
 					}
 
-					txRef, err := idx.GetTxHashReference(tx.TxID[:])
+					txRef, err := idx.GetTxMetadataByHash(tx.TxID[:])
 					qt.Check(b, err, qt.IsNil)
 					if err == nil {
 						qt.Check(b, txRef.BlockHeight, qt.Equals, vote.Height)
@@ -138,7 +141,11 @@ func BenchmarkFetchTx(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		for j := 0; j < numTxs; j++ {
-			idx.OnNewTx(&vochaintx.Tx{TxID: util.Random32()}, uint32(i), int32(j))
+			idx.OnNewTx(&vochaintx.Tx{
+				TxID:        util.Random32(),
+				TxModelType: "vote",
+				Tx:          &models.Tx{Payload: &models.Tx_Vote{}},
+			}, uint32(i), int32(j))
 		}
 		err := idx.Commit(uint32(i))
 		qt.Assert(b, err, qt.IsNil)
@@ -147,14 +154,14 @@ func BenchmarkFetchTx(b *testing.B) {
 
 		startTime := time.Now()
 		for j := 0; j < numTxs; j++ {
-			_, err = idx.GetTransaction(uint64((i * numTxs) + j + 1))
+			_, err = idx.GetTransactionByHeightAndIndex(int64(i), int64(j))
 			qt.Assert(b, err, qt.IsNil)
 		}
-		log.Infof("fetched %d transactions (out of %d total) by index, took %s",
+		log.Infof("fetched %d transactions (out of %d total) by height+index, took %s",
 			numTxs, (i+1)*numTxs, time.Since(startTime))
 		startTime = time.Now()
 		for j := 0; j < numTxs; j++ {
-			_, err = idx.GetTxHashReference([]byte(fmt.Sprintf("hash%d%d", i, j)))
+			_, err = idx.GetTxMetadataByHash([]byte(fmt.Sprintf("hash%d%d", i, j)))
 			qt.Assert(b, err, qt.IsNil)
 		}
 		log.Infof("fetched %d transactions (out of %d total) by hash, took %s",
@@ -191,4 +198,85 @@ func BenchmarkNewProcess(b *testing.B) {
 	app.AdvanceTestBlock()
 	log.Infof("indexed %d new processes, took %s",
 		numProcesses, time.Since(startTime))
+}
+
+func BenchmarkBlockList(b *testing.B) {
+	app := vochain.TestBaseApplication(b)
+
+	idx, err := New(app, Options{DataDir: b.TempDir()})
+	qt.Assert(b, err, qt.IsNil)
+
+	count := 100000
+
+	createDummyBlocks(b, idx, count)
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	benchmarkBlockList := func(b *testing.B,
+		limit int, offset int, chainID string, hash string, proposerAddress string,
+	) {
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			blocks, total, err := idx.BlockList(limit, offset, chainID, hash, proposerAddress)
+			qt.Assert(b, err, qt.IsNil)
+			qt.Assert(b, blocks, qt.HasLen, limit, qt.Commentf("%+v", blocks))
+			qt.Assert(b, blocks[0].TxCount, qt.Equals, int64(0))
+			qt.Assert(b, total, qt.Equals, uint64(count))
+		}
+	}
+
+	// Run sub-benchmarks with different limits and filters
+	b.Run("BlockListLimit1", func(b *testing.B) {
+		benchmarkBlockList(b, 1, 0, "", "", "")
+	})
+
+	b.Run("BlockListLimit10", func(b *testing.B) {
+		benchmarkBlockList(b, 10, 0, "", "", "")
+	})
+
+	b.Run("BlockListLimit100", func(b *testing.B) {
+		benchmarkBlockList(b, 100, 0, "", "", "")
+	})
+
+	b.Run("BlockListOffset", func(b *testing.B) {
+		benchmarkBlockList(b, 10, count/2, "", "", "")
+	})
+
+	b.Run("BlockListWithChainID", func(b *testing.B) {
+		benchmarkBlockList(b, 10, 0, "test", "", "")
+	})
+
+	b.Run("BlockListWithHashSubstr", func(b *testing.B) {
+		benchmarkBlockList(b, 10, 0, "", "cafe", "")
+	})
+	b.Run("BlockListWithHashExact", func(b *testing.B) {
+		benchmarkBlockList(b, 10, 0, "", "cafecafecafecafecafecafecafecafecafecafecafecafecafecafecafecafe", "")
+	})
+}
+
+func createDummyBlocks(b *testing.B, idx *Indexer, n int) {
+	idx.blockMu.Lock()
+	defer idx.blockMu.Unlock()
+
+	queries := idx.blockTxQueries()
+	for h := 1; h <= n; h++ {
+		_, err := queries.CreateBlock(context.TODO(), indexerdb.CreateBlockParams{
+			ChainID: "test",
+			Height:  int64(h),
+			Time:    time.Now(),
+			Hash: nonNullBytes([]byte{
+				0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe,
+				0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe, 0xca, 0xfe,
+			}),
+			ProposerAddress: nonNullBytes([]byte{0xfe, 0xde}),
+			LastBlockHash:   nonNullBytes([]byte{0xca, 0xfe}),
+		},
+		)
+		qt.Assert(b, err, qt.IsNil)
+	}
+	err := idx.blockTx.Commit()
+	qt.Assert(b, err, qt.IsNil)
 }
