@@ -30,11 +30,17 @@ func Prepare(ctx context.Context, db DBTX) (*Queries, error) {
 	if q.countAccountsStmt, err = db.PrepareContext(ctx, countAccounts); err != nil {
 		return nil, fmt.Errorf("error preparing query CountAccounts: %w", err)
 	}
+	if q.countBlocksStmt, err = db.PrepareContext(ctx, countBlocks); err != nil {
+		return nil, fmt.Errorf("error preparing query CountBlocks: %w", err)
+	}
 	if q.countTokenTransfersByAccountStmt, err = db.PrepareContext(ctx, countTokenTransfersByAccount); err != nil {
 		return nil, fmt.Errorf("error preparing query CountTokenTransfersByAccount: %w", err)
 	}
 	if q.countTransactionsStmt, err = db.PrepareContext(ctx, countTransactions); err != nil {
 		return nil, fmt.Errorf("error preparing query CountTransactions: %w", err)
+	}
+	if q.countTransactionsByHeightStmt, err = db.PrepareContext(ctx, countTransactionsByHeight); err != nil {
+		return nil, fmt.Errorf("error preparing query CountTransactionsByHeight: %w", err)
 	}
 	if q.countVotesStmt, err = db.PrepareContext(ctx, countVotes); err != nil {
 		return nil, fmt.Errorf("error preparing query CountVotes: %w", err)
@@ -60,8 +66,11 @@ func Prepare(ctx context.Context, db DBTX) (*Queries, error) {
 	if q.createVoteStmt, err = db.PrepareContext(ctx, createVote); err != nil {
 		return nil, fmt.Errorf("error preparing query CreateVote: %w", err)
 	}
-	if q.getBlockStmt, err = db.PrepareContext(ctx, getBlock); err != nil {
-		return nil, fmt.Errorf("error preparing query GetBlock: %w", err)
+	if q.getBlockByHashStmt, err = db.PrepareContext(ctx, getBlockByHash); err != nil {
+		return nil, fmt.Errorf("error preparing query GetBlockByHash: %w", err)
+	}
+	if q.getBlockByHeightStmt, err = db.PrepareContext(ctx, getBlockByHeight); err != nil {
+		return nil, fmt.Errorf("error preparing query GetBlockByHeight: %w", err)
 	}
 	if q.getEntityCountStmt, err = db.PrepareContext(ctx, getEntityCount); err != nil {
 		return nil, fmt.Errorf("error preparing query GetEntityCount: %w", err)
@@ -81,20 +90,23 @@ func Prepare(ctx context.Context, db DBTX) (*Queries, error) {
 	if q.getTokenTransferStmt, err = db.PrepareContext(ctx, getTokenTransfer); err != nil {
 		return nil, fmt.Errorf("error preparing query GetTokenTransfer: %w", err)
 	}
-	if q.getTransactionStmt, err = db.PrepareContext(ctx, getTransaction); err != nil {
-		return nil, fmt.Errorf("error preparing query GetTransaction: %w", err)
-	}
 	if q.getTransactionByHashStmt, err = db.PrepareContext(ctx, getTransactionByHash); err != nil {
 		return nil, fmt.Errorf("error preparing query GetTransactionByHash: %w", err)
 	}
-	if q.getTxReferenceByBlockHeightAndBlockIndexStmt, err = db.PrepareContext(ctx, getTxReferenceByBlockHeightAndBlockIndex); err != nil {
-		return nil, fmt.Errorf("error preparing query GetTxReferenceByBlockHeightAndBlockIndex: %w", err)
+	if q.getTransactionByHeightAndIndexStmt, err = db.PrepareContext(ctx, getTransactionByHeightAndIndex); err != nil {
+		return nil, fmt.Errorf("error preparing query GetTransactionByHeightAndIndex: %w", err)
 	}
 	if q.getVoteStmt, err = db.PrepareContext(ctx, getVote); err != nil {
 		return nil, fmt.Errorf("error preparing query GetVote: %w", err)
 	}
+	if q.lastBlockHeightStmt, err = db.PrepareContext(ctx, lastBlockHeight); err != nil {
+		return nil, fmt.Errorf("error preparing query LastBlockHeight: %w", err)
+	}
 	if q.searchAccountsStmt, err = db.PrepareContext(ctx, searchAccounts); err != nil {
 		return nil, fmt.Errorf("error preparing query SearchAccounts: %w", err)
+	}
+	if q.searchBlocksStmt, err = db.PrepareContext(ctx, searchBlocks); err != nil {
+		return nil, fmt.Errorf("error preparing query SearchBlocks: %w", err)
 	}
 	if q.searchEntitiesStmt, err = db.PrepareContext(ctx, searchEntities); err != nil {
 		return nil, fmt.Errorf("error preparing query SearchEntities: %w", err)
@@ -147,6 +159,11 @@ func (q *Queries) Close() error {
 			err = fmt.Errorf("error closing countAccountsStmt: %w", cerr)
 		}
 	}
+	if q.countBlocksStmt != nil {
+		if cerr := q.countBlocksStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing countBlocksStmt: %w", cerr)
+		}
+	}
 	if q.countTokenTransfersByAccountStmt != nil {
 		if cerr := q.countTokenTransfersByAccountStmt.Close(); cerr != nil {
 			err = fmt.Errorf("error closing countTokenTransfersByAccountStmt: %w", cerr)
@@ -155,6 +172,11 @@ func (q *Queries) Close() error {
 	if q.countTransactionsStmt != nil {
 		if cerr := q.countTransactionsStmt.Close(); cerr != nil {
 			err = fmt.Errorf("error closing countTransactionsStmt: %w", cerr)
+		}
+	}
+	if q.countTransactionsByHeightStmt != nil {
+		if cerr := q.countTransactionsByHeightStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing countTransactionsByHeightStmt: %w", cerr)
 		}
 	}
 	if q.countVotesStmt != nil {
@@ -197,9 +219,14 @@ func (q *Queries) Close() error {
 			err = fmt.Errorf("error closing createVoteStmt: %w", cerr)
 		}
 	}
-	if q.getBlockStmt != nil {
-		if cerr := q.getBlockStmt.Close(); cerr != nil {
-			err = fmt.Errorf("error closing getBlockStmt: %w", cerr)
+	if q.getBlockByHashStmt != nil {
+		if cerr := q.getBlockByHashStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing getBlockByHashStmt: %w", cerr)
+		}
+	}
+	if q.getBlockByHeightStmt != nil {
+		if cerr := q.getBlockByHeightStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing getBlockByHeightStmt: %w", cerr)
 		}
 	}
 	if q.getEntityCountStmt != nil {
@@ -232,19 +259,14 @@ func (q *Queries) Close() error {
 			err = fmt.Errorf("error closing getTokenTransferStmt: %w", cerr)
 		}
 	}
-	if q.getTransactionStmt != nil {
-		if cerr := q.getTransactionStmt.Close(); cerr != nil {
-			err = fmt.Errorf("error closing getTransactionStmt: %w", cerr)
-		}
-	}
 	if q.getTransactionByHashStmt != nil {
 		if cerr := q.getTransactionByHashStmt.Close(); cerr != nil {
 			err = fmt.Errorf("error closing getTransactionByHashStmt: %w", cerr)
 		}
 	}
-	if q.getTxReferenceByBlockHeightAndBlockIndexStmt != nil {
-		if cerr := q.getTxReferenceByBlockHeightAndBlockIndexStmt.Close(); cerr != nil {
-			err = fmt.Errorf("error closing getTxReferenceByBlockHeightAndBlockIndexStmt: %w", cerr)
+	if q.getTransactionByHeightAndIndexStmt != nil {
+		if cerr := q.getTransactionByHeightAndIndexStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing getTransactionByHeightAndIndexStmt: %w", cerr)
 		}
 	}
 	if q.getVoteStmt != nil {
@@ -252,9 +274,19 @@ func (q *Queries) Close() error {
 			err = fmt.Errorf("error closing getVoteStmt: %w", cerr)
 		}
 	}
+	if q.lastBlockHeightStmt != nil {
+		if cerr := q.lastBlockHeightStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing lastBlockHeightStmt: %w", cerr)
+		}
+	}
 	if q.searchAccountsStmt != nil {
 		if cerr := q.searchAccountsStmt.Close(); cerr != nil {
 			err = fmt.Errorf("error closing searchAccountsStmt: %w", cerr)
+		}
+	}
+	if q.searchBlocksStmt != nil {
+		if cerr := q.searchBlocksStmt.Close(); cerr != nil {
+			err = fmt.Errorf("error closing searchBlocksStmt: %w", cerr)
 		}
 	}
 	if q.searchEntitiesStmt != nil {
@@ -354,85 +386,93 @@ func (q *Queries) queryRow(ctx context.Context, stmt *sql.Stmt, query string, ar
 }
 
 type Queries struct {
-	db                                           DBTX
-	tx                                           *sql.Tx
-	computeProcessVoteCountStmt                  *sql.Stmt
-	countAccountsStmt                            *sql.Stmt
-	countTokenTransfersByAccountStmt             *sql.Stmt
-	countTransactionsStmt                        *sql.Stmt
-	countVotesStmt                               *sql.Stmt
-	createAccountStmt                            *sql.Stmt
-	createBlockStmt                              *sql.Stmt
-	createProcessStmt                            *sql.Stmt
-	createTokenFeeStmt                           *sql.Stmt
-	createTokenTransferStmt                      *sql.Stmt
-	createTransactionStmt                        *sql.Stmt
-	createVoteStmt                               *sql.Stmt
-	getBlockStmt                                 *sql.Stmt
-	getEntityCountStmt                           *sql.Stmt
-	getProcessStmt                               *sql.Stmt
-	getProcessCountStmt                          *sql.Stmt
-	getProcessIDsByFinalResultsStmt              *sql.Stmt
-	getProcessStatusStmt                         *sql.Stmt
-	getTokenTransferStmt                         *sql.Stmt
-	getTransactionStmt                           *sql.Stmt
-	getTransactionByHashStmt                     *sql.Stmt
-	getTxReferenceByBlockHeightAndBlockIndexStmt *sql.Stmt
-	getVoteStmt                                  *sql.Stmt
-	searchAccountsStmt                           *sql.Stmt
-	searchEntitiesStmt                           *sql.Stmt
-	searchProcessesStmt                          *sql.Stmt
-	searchTokenFeesStmt                          *sql.Stmt
-	searchTokenTransfersStmt                     *sql.Stmt
-	searchTransactionsStmt                       *sql.Stmt
-	searchVotesStmt                              *sql.Stmt
-	setProcessResultsCancelledStmt               *sql.Stmt
-	setProcessResultsReadyStmt                   *sql.Stmt
-	updateProcessEndDateStmt                     *sql.Stmt
-	updateProcessFromStateStmt                   *sql.Stmt
-	updateProcessResultByIDStmt                  *sql.Stmt
-	updateProcessResultsStmt                     *sql.Stmt
+	db                                 DBTX
+	tx                                 *sql.Tx
+	computeProcessVoteCountStmt        *sql.Stmt
+	countAccountsStmt                  *sql.Stmt
+	countBlocksStmt                    *sql.Stmt
+	countTokenTransfersByAccountStmt   *sql.Stmt
+	countTransactionsStmt              *sql.Stmt
+	countTransactionsByHeightStmt      *sql.Stmt
+	countVotesStmt                     *sql.Stmt
+	createAccountStmt                  *sql.Stmt
+	createBlockStmt                    *sql.Stmt
+	createProcessStmt                  *sql.Stmt
+	createTokenFeeStmt                 *sql.Stmt
+	createTokenTransferStmt            *sql.Stmt
+	createTransactionStmt              *sql.Stmt
+	createVoteStmt                     *sql.Stmt
+	getBlockByHashStmt                 *sql.Stmt
+	getBlockByHeightStmt               *sql.Stmt
+	getEntityCountStmt                 *sql.Stmt
+	getProcessStmt                     *sql.Stmt
+	getProcessCountStmt                *sql.Stmt
+	getProcessIDsByFinalResultsStmt    *sql.Stmt
+	getProcessStatusStmt               *sql.Stmt
+	getTokenTransferStmt               *sql.Stmt
+	getTransactionByHashStmt           *sql.Stmt
+	getTransactionByHeightAndIndexStmt *sql.Stmt
+	getVoteStmt                        *sql.Stmt
+	lastBlockHeightStmt                *sql.Stmt
+	searchAccountsStmt                 *sql.Stmt
+	searchBlocksStmt                   *sql.Stmt
+	searchEntitiesStmt                 *sql.Stmt
+	searchProcessesStmt                *sql.Stmt
+	searchTokenFeesStmt                *sql.Stmt
+	searchTokenTransfersStmt           *sql.Stmt
+	searchTransactionsStmt             *sql.Stmt
+	searchVotesStmt                    *sql.Stmt
+	setProcessResultsCancelledStmt     *sql.Stmt
+	setProcessResultsReadyStmt         *sql.Stmt
+	updateProcessEndDateStmt           *sql.Stmt
+	updateProcessFromStateStmt         *sql.Stmt
+	updateProcessResultByIDStmt        *sql.Stmt
+	updateProcessResultsStmt           *sql.Stmt
 }
 
 func (q *Queries) WithTx(tx *sql.Tx) *Queries {
 	return &Queries{
-		db:                               tx,
-		tx:                               tx,
-		computeProcessVoteCountStmt:      q.computeProcessVoteCountStmt,
-		countAccountsStmt:                q.countAccountsStmt,
-		countTokenTransfersByAccountStmt: q.countTokenTransfersByAccountStmt,
-		countTransactionsStmt:            q.countTransactionsStmt,
-		countVotesStmt:                   q.countVotesStmt,
-		createAccountStmt:                q.createAccountStmt,
-		createBlockStmt:                  q.createBlockStmt,
-		createProcessStmt:                q.createProcessStmt,
-		createTokenFeeStmt:               q.createTokenFeeStmt,
-		createTokenTransferStmt:          q.createTokenTransferStmt,
-		createTransactionStmt:            q.createTransactionStmt,
-		createVoteStmt:                   q.createVoteStmt,
-		getBlockStmt:                     q.getBlockStmt,
-		getEntityCountStmt:               q.getEntityCountStmt,
-		getProcessStmt:                   q.getProcessStmt,
-		getProcessCountStmt:              q.getProcessCountStmt,
-		getProcessIDsByFinalResultsStmt:  q.getProcessIDsByFinalResultsStmt,
-		getProcessStatusStmt:             q.getProcessStatusStmt,
-		getTokenTransferStmt:             q.getTokenTransferStmt,
-		getTransactionStmt:               q.getTransactionStmt,
-		getTransactionByHashStmt:         q.getTransactionByHashStmt,
-		getTxReferenceByBlockHeightAndBlockIndexStmt: q.getTxReferenceByBlockHeightAndBlockIndexStmt,
-		getVoteStmt:                    q.getVoteStmt,
-		searchAccountsStmt:             q.searchAccountsStmt,
-		searchEntitiesStmt:             q.searchEntitiesStmt,
-		searchProcessesStmt:            q.searchProcessesStmt,
-		searchTokenFeesStmt:            q.searchTokenFeesStmt,
-		searchTokenTransfersStmt:       q.searchTokenTransfersStmt,
-		searchTransactionsStmt:         q.searchTransactionsStmt,
-		searchVotesStmt:                q.searchVotesStmt,
-		setProcessResultsCancelledStmt: q.setProcessResultsCancelledStmt,
-		setProcessResultsReadyStmt:     q.setProcessResultsReadyStmt,
-		updateProcessEndDateStmt:       q.updateProcessEndDateStmt,
-		updateProcessFromStateStmt:     q.updateProcessFromStateStmt,
-		updateProcessResultByIDStmt:    q.updateProcessResultByIDStmt,
-		updateProcessResultsStmt:       q.updateProcessResultsStmt,
+		db:                                 tx,
+		tx:                                 tx,
+		computeProcessVoteCountStmt:        q.computeProcessVoteCountStmt,
+		countAccountsStmt:                  q.countAccountsStmt,
+		countBlocksStmt:                    q.countBlocksStmt,
+		countTokenTransfersByAccountStmt:   q.countTokenTransfersByAccountStmt,
+		countTransactionsStmt:              q.countTransactionsStmt,
+		countTransactionsByHeightStmt:      q.countTransactionsByHeightStmt,
+		countVotesStmt:                     q.countVotesStmt,
+		createAccountStmt:                  q.createAccountStmt,
+		createBlockStmt:                    q.createBlockStmt,
+		createProcessStmt:                  q.createProcessStmt,
+		createTokenFeeStmt:                 q.createTokenFeeStmt,
+		createTokenTransferStmt:            q.createTokenTransferStmt,
+		createTransactionStmt:              q.createTransactionStmt,
+		createVoteStmt:                     q.createVoteStmt,
+		getBlockByHashStmt:                 q.getBlockByHashStmt,
+		getBlockByHeightStmt:               q.getBlockByHeightStmt,
+		getEntityCountStmt:                 q.getEntityCountStmt,
+		getProcessStmt:                     q.getProcessStmt,
+		getProcessCountStmt:                q.getProcessCountStmt,
+		getProcessIDsByFinalResultsStmt:    q.getProcessIDsByFinalResultsStmt,
+		getProcessStatusStmt:               q.getProcessStatusStmt,
+		getTokenTransferStmt:               q.getTokenTransferStmt,
+		getTransactionByHashStmt:           q.getTransactionByHashStmt,
+		getTransactionByHeightAndIndexStmt: q.getTransactionByHeightAndIndexStmt,
+		getVoteStmt:                        q.getVoteStmt,
+		lastBlockHeightStmt:                q.lastBlockHeightStmt,
+		searchAccountsStmt:                 q.searchAccountsStmt,
+		searchBlocksStmt:                   q.searchBlocksStmt,
+		searchEntitiesStmt:                 q.searchEntitiesStmt,
+		searchProcessesStmt:                q.searchProcessesStmt,
+		searchTokenFeesStmt:                q.searchTokenFeesStmt,
+		searchTokenTransfersStmt:           q.searchTokenTransfersStmt,
+		searchTransactionsStmt:             q.searchTransactionsStmt,
+		searchVotesStmt:                    q.searchVotesStmt,
+		setProcessResultsCancelledStmt:     q.setProcessResultsCancelledStmt,
+		setProcessResultsReadyStmt:         q.setProcessResultsReadyStmt,
+		updateProcessEndDateStmt:           q.updateProcessEndDateStmt,
+		updateProcessFromStateStmt:         q.updateProcessFromStateStmt,
+		updateProcessResultByIDStmt:        q.updateProcessResultByIDStmt,
+		updateProcessResultsStmt:           q.updateProcessResultsStmt,
 	}
 }
